add router hooks module
