+++ frontend/src/App.vue
@@ -169,8 +169,7 @@ import { CzNotifications, Notifications } from "@cznethub/cznet-vue-core";
 import { Subscription } from "rxjs";
 import User from "@/models/user.model";
 import CdLogin from "@/components/account/cd.login.vue";
-import { RouteLocationRaw, useRoute, useRouter } from "vue-router";
-import { addRouteTags } from "./guards";
+import { RouteLocationRaw, useRoute } from "vue-router";
 
 @Component({
   name: "app",
@@ -242,9 +241,6 @@ class App extends Vue {
 
   async created() {
     document.title = APP_NAME;
-
-    const router = useRouter();
-    router.afterEach(addRouteTags);
 
     User.fetchSchemas();
     // Guards are setup after checking authorization and loading access tokens
